Update misc/rpm to rpm-4.9.1.1
rpm-4.9.1.1 adds a lot, but it's actually compatible with RPMs in the
wild now.  Marc Espie said he did not mind me becoming the maintainer
of misc/rpm as he does not regularly use it and I do.  I do not have
an openbsd.org account, so I've just put my personal address in there
for now.

It builds and packages cleanly for me on amd64 -current.  I removed
the pre-configure: target I had before and make use of the 'autoconf'
CONFIGURE_STYLE feature in the ports system.  Seems to do the trick.