Fix 'regress' target for misc/rpm, but still working on test cases.
Hooked up the regress target for misc/rpm and patched the driver so
it works with /bin/sh on OpenBSD and does not require bash.  However,
there are a number of test case failures that I am still working on
resolving.  I've added the regress.log here so people can see what
is failing for now.