whoops I thought I was committing in all related ports... anyway previous commit message follows...
update and enforce the correct way of having versions. Found the hard way by trying to enforce a productivity/qzeitgeist version, and after experimenting and conforming, hunt and fix in KDE related ports. This has to be fixed in another port in openbsd-wip:multimedia/xmltv and maybe others in openbsd tree....