Workaround for the bug in the docs of racket.
Due to a bug, racket generates different files on amd64 and i386, so a
single PLIST for all the architectures will break the generation of the
package.

I'm creating a PFRAG file for each architecture with the broken
directory. This commit contains the amd64 part, I'll commit the i386
file this weekend.