Added a port for the_silver_searcher a fast grep like utility with some additional features.
The licence is Apache 2 which seems BSD compatible, yet the build was
not a straight forward. The source code does not ship its own configure
but you have to bootstrap it. And that was far from simple: I ended up
patching the provided build.sh to make the_silver_searcher work. This is
definitely something I have to contact the developer about.