Import libscrypt, a library for using scrypt.
This includes a fix from upstream that is not present in the release
package, and a replacement of the base64 en/decode implementation.
The original base64 stuff is not compatible with strict alignment
architectures.

This is for testing, and playing around with, until it will be clear how
upstream wants to handle the lazy base64 implementation.