Import x11/srandr
srandrd is a simple daemon that executes a command when the xorg output
changes, i.e. a monitor is plugged or unplugged. By default srandrd
forks to background and exits when the xserver exits.

WIP:
Not working for me yet. After i change my monitor configuration,
i have to call `xrandr --auto` to trigger the change notification.