add new port nomacs 2.2.0
nomacs is small, fast and able to handle the most common image
formats. Additionally it is possible to synchronize multiple
viewers. A synchronization of viewers running on the same computer
or via LAN is possible. It allows to compare images and spot the
differences (e.g. schemes of architects to show the progress).