graphics/ffmpeg: isolated fdk-aac support in "gpl3" flavor
As pointed out by brad@, fdk-aac (and other AAC encoders and decoders) are
distributed under Apache 2 license, which is incompatible with GPLv2, but
compatible with GPLv3.

While GPLv3 is a show-stopper for some, it is not a big deal for others, so
isolating GPLv2 incompatible stuff in "gpl3" flavor seems to be a viable
compromise between functionality and legal sanity.