www/edbrowse: new port
This program is, at first glance, a reimplementation of /bin/ed.  In fact you
might issue a few ed commands and not realize that you are actually running this
program.  But as you proceed you will eventually discover some discrepancies,
areas where edbrowse differs from ed.  These are discussed below.

Reinventing ed seems like a complete waste of time, until you realize that this
program also acts as a browser - a browser embedded inside ed.  You can edit a
URL as easily as a local file, and activate browse mode to render the html tags
in a manner that is appropriate for a command-response program such as this.  In
other words, we discard most of the formatting information and retain the links
and fill-out forms.  This allows blind  users to access the Internet through an
application that is compatible with the linear nature of speech or braille.