editors/vis: vim-like text editor
Vis aims to be a modern, legacy free, simple yet efficient vim-like editor.
It extends vim's modal editing with built-in support for multiple
cursors/selections and combines it with sam's structural regular expression
based command language.  As an universal editor it has decent Unicode support
(including double width and combining characters) and should cope with arbitrary
files.  Efficient syntax highlighting is provided using Parsing Expression
Grammars which can be conveniently expressed using Lua in form of LPeg.

The editor core is written in a reasonable amount of clean (your mileage may
vary), modern and legacy free C code enabling it to run in resource constrained
environments. The implementation should be easy to hack on and encourage
experimentation (e.g. native built in support for multiple cursors). There also
exists a Lua API for in process extensions.

Vis strives to be simple and focuses on its core task: efficient text
management. As an example the file open dialog is provided by an independent
utility. There exist plans to use a client/server architecture, delegating
window management to your windowing system or favorite terminal multiplexer.

The intention is not to be bug for bug compatible with vim, instead a similar
editing experience should be provided. The goal could thus be summarized as "80%
of vim's features implemented in roughly 1% of the code".