Remove emulators/vecx
I have no interest to work on this further at the moment. ROM path is
hardcoded, and even when patching to use a specific directory, we still
need to ask the user to put it there manually as we cannot distribute
the ROM. Not very user friendly.

For the time being MESS is a superior alternative.