change how xash3d and associated game logic packages are handled. instead of installing xash3d with a script that expects half-life game logic libraries pre-installed when they aren't, make the xash3d script instead refer to the extra packages to support the games.
the xash3d-hl package has been updated to include a `halflife' script
for launching the game.  this is done to prepare for also supporting
a xash3d-bshift package, for Half-Life: Blue Shift.

also reverted the version number from 2.3 on the game logic to
simply 0.0.${gitcommitid}, as 2.3 isn't really the truth with all
the updates and changes from upstream.