Add some || defined(__OpenBSD__) patches
Some of these are probably wrong, and to upstream they should probably
by reversed to "#if defined(__APPLE__) || defined(__WIN32__)" or
somesuch.  Some need to be defining features in some main header and
reusing _that_ for clarity.

At least with this it "starts", although lots of "window does not exist"
errors.