Add lang/lfortran.
LFortran is a modern open-source (BSD licensed) interactive Fortran
compiler built on top of LLVM. It can execute user's code interactively
to allow exploratory work (much like Python, MATLAB or Julia) as well
as compile to binaries with the goal to run user's code on modern
architectures such as multi-core CPUs and GPUs.

LFortran is in alpha (it is expected to not work on third-party codes
and users enthusiastically participate in bug reporting and fixing).